// frontend/projects/partner/src/app/api/model/contact-person-dto.ts
/**
 * BDK-CoreFinance API Gateway
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {AddressDto} from './address-dto';
import {PartnerFunctionDto} from './partner-function-dto';


/**
 * Contact person information containing personal and business details
 */
export interface ContactPersonDto {
  /**
   * Unique person identifier - primary key. Optional for new contact persons (auto-generated), required for updates.
   */
  personId?: number;
  /**
   * Partner function identifier
   */
  partnerFunctionId: number;
  /**
   * Associated partner identifier
   */
  partnerNumber?: number;
  /**
   * Electronic Data Processing name/identifier
   */
  nameEdp?: string;
  /**
   * Name part 1 - typically surname or company name (required)
   */
  name1: string;
  /**
   * Name part 2 - additional name information
   */
  name2?: string;
  /**
   * Name part 3 - further name information
   */
  name3?: string;
  /**
   * Alpha code identifier - short name (required)
   */
  alphaCode: string;
  /**
   * Date of birth
   */
  dateOfBirth?: string;
  /**
   * Department identifier
   */
  departmentId?: number;
  /**
   * Department name
   */
  departmentName?: string;
  /**
   * Hobby identifier
   */
  hobbyId?: number;
  /**
   * Origin type code
   */
  originType?: string;
  /**
   * Origin number
   */
  originNumber?: number;
  /**
   * Gender code (1=male, 2=female)
   */
  gender?: number;
  /**
   * Last modification date
   */
  modificationDate?: string;
  /**
   * ID of person who made last modification
   */
  modifierId?: string;
  /**
   * Credit contract flag
   */
  creditContract?: boolean;
  /**
   * Leasing contract flag
   */
  leasingContract?: boolean;
  /**
   * Leasing payer flag
   */
  leasingPayer?: boolean;
  /**
   * Salutation code
   */
  salutationCode?: number;
  /**
   * Letter salutation text
   */
  letterSalutation?: string;
  /**
   * Academic title
   */
  title?: string;
  /**
   * First name
   */
  firstName?: string;
  /**
   * Identification line
   */
  identificationLine?: string;
  /**
   * Marketing stop flag
   */
  marketingStop?: boolean;
  /**
   * Format CO flag
   */
  formatCo?: boolean;
  /**
   * Secondary person identifier
   */
  personId2?: number;
  /**
   * Database number 1
   */
  dbNumber1?: number;
  /**
   * Database number 2
   */
  dbNumber2?: number;
  /**
   * Database number 3
   */
  dbNumber3?: number;
  /**
   * WS reason/cause
   */
  wsReason?: string;
  partnerFunction?: PartnerFunctionDto;
  /**
   * List of addresses associated with this contact person
   */
  addresses?: Array<AddressDto>;
}